move copyUtils and walk to 'utils' folder
