+++ Composer/packages/lib/shared/src/utils/walkAdaptiveAction.ts
